+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Validate Cilium CNI.
+- Validate `cilium.giantswarm.io/pod-cidr` annotation is present and valid while upgrading from v17 to v18.
 
 ## [4.1.0] - 2022-07-22