+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.2.0] - 2022-08-04
+
 ### Added
 
 - Validate `cilium.giantswarm.io/pod-cidr` annotation does not intersect with current CIDR's.
@@ -351,7 +353,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/aws-admission-controller/compare/v4.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-admission-controller/compare/v4.2.0...HEAD
+[4.2.0]: https://github.com/giantswarm/aws-admission-controller/compare/v4.1.0...v4.2.0
 [4.1.0]: https://github.com/giantswarm/aws-admission-controller/compare/v4.0.3...v4.1.0
 [4.0.3]: https://github.com/giantswarm/aws-admission-controller/compare/v4.0.2...v4.0.3
 [4.0.2]: https://github.com/giantswarm/aws-admission-controller/compare/v4.0.1...v4.0.2